enable Eclipse Null Analysis

+++ plugin/src/main/java/de/sebthom/eclipse/open_with_eclipse/WindowsRegistryHelper.java
@@ -9,7 +9,7 @@ import static de.sebthom.eclipse.commons.system.win.WindowsRegistry.*;
 import java.nio.file.Files;
 import java.nio.file.Path;
 
-import org.eclipse.jdt.annotation.NonNull;
+import org.eclipse.jdt.annotation.Nullable;
 
 import de.sebthom.eclipse.commons.system.win.WindowsRegistry;
 import net.sf.jstuff.core.Strings;
@@ -21,9 +21,9 @@ import net.sf.jstuff.core.validation.Args;
 public abstract class WindowsRegistryHelper {
 
    private static final String SOFTWARE_CLASSES_WILDCARD_SHELL = WindowsRegistry.join("SOFTWARE", "Classes", "*", "shell");
-   private static final String DEFAULT_VALUE_NAME = null;
+   private static final @Nullable String DEFAULT_VALUE_NAME = null;
 
-   public static void createOpenWithRegistryEntry(@NonNull final Path launcherExe, @NonNull final String contextMenuEntryLabel) {
+   public static void createOpenWithRegistryEntry(final Path launcherExe, final String contextMenuEntryLabel) {
       Args.isFileReadable("launcherExe", launcherExe);
       Args.notBlank("contextMenuEntryLabel", contextMenuEntryLabel);
 
@@ -34,14 +34,14 @@ public abstract class WindowsRegistryHelper {
       HKEY_CURRENT_USER.setStringValue(keyPath, "command", DEFAULT_VALUE_NAME, getOpenWithCommand(launcherExe));
    }
 
-   private static String getOpenWithCommand(@NonNull final Path launcherExe) {
+   private static String getOpenWithCommand(final Path launcherExe) {
       return "\"" + launcherExe.toAbsolutePath() + "\" \"%V\"";
    }
 
    /**
     * @return a corresponding Open with entry from the registry for the given Eclipse exe or null
     */
-   private static String getOpenWithRegistryEntry(@NonNull final Path launcherExe) {
+   private static @Nullable String getOpenWithRegistryEntry(final Path launcherExe) {
       final var openWithCommand = getOpenWithCommand(launcherExe);
 
       for (final var keyName : HKEY_CURRENT_USER.getKeys(SOFTWARE_CLASSES_WILDCARD_SHELL)) {
@@ -56,7 +56,7 @@ public abstract class WindowsRegistryHelper {
       return hasOpenWithRegistryEntry(Constants.LAUNCHER_EXE);
    }
 
-   private static boolean hasOpenWithRegistryEntry(@NonNull final Path launcherExe) {
+   private static boolean hasOpenWithRegistryEntry(final Path launcherExe) {
       Args.notNull("launcherExe", launcherExe);
 
       return getOpenWithRegistryEntry(launcherExe) != null;
@@ -69,18 +69,20 @@ public abstract class WindowsRegistryHelper {
          final var entryCreatedBy = HKEY_CURRENT_USER.getStringValue(keyPath, "entryCreatedBy");
          if (Plugin.PLUGIN_ID.equals(entryCreatedBy)) {
             final var command = HKEY_CURRENT_USER.getStringValue(keyPath, "command", DEFAULT_VALUE_NAME);
-            final var exePath = Strings.substringBetween(command, "\"", "\"");
-            if (Strings.endsWithIgnoreCase(exePath, ".exe") //
-               && !Files.exists(Path.of(exePath))) {
-               HKEY_CURRENT_USER.deleteKey(keyPath, true);
-               deleted++;
+            if (command != null) {
+               final var exePath = Strings.substringBetween(command, "\"", "\"");
+               if (Strings.endsWithIgnoreCase(exePath, ".exe") //
+                  && !Files.exists(Path.of(exePath))) {
+                  HKEY_CURRENT_USER.deleteKey(keyPath, true);
+                  deleted++;
+               }
             }
          }
       }
       return deleted;
    }
 
-   public static int removeOpenWithRegistryEntries(@NonNull final Path launcherExe) {
+   public static int removeOpenWithRegistryEntries(final Path launcherExe) {
       var deleted = 0;
       final var launcherExePath = launcherExe.toAbsolutePath().toString();
       for (final var keyName : HKEY_CURRENT_USER.getKeys(SOFTWARE_CLASSES_WILDCARD_SHELL)) {
@@ -88,10 +90,12 @@ public abstract class WindowsRegistryHelper {
          final var entryCreatedBy = HKEY_CURRENT_USER.getStringValue(keyPath, "entryCreatedBy");
          if (Plugin.PLUGIN_ID.equals(entryCreatedBy)) {
             final var command = HKEY_CURRENT_USER.getStringValue(keyPath, "command", DEFAULT_VALUE_NAME);
-            final var exePath = Strings.substringBetween(command, "\"", "\"");
-            if (launcherExePath.equalsIgnoreCase(exePath)) {
-               HKEY_CURRENT_USER.deleteKey(keyPath, true);
-               deleted++;
+            if (command != null) {
+               final var exePath = Strings.substringBetween(command, "\"", "\"");
+               if (launcherExePath.equalsIgnoreCase(exePath)) {
+                  HKEY_CURRENT_USER.deleteKey(keyPath, true);
+                  deleted++;
+               }
             }
          }
       }
